refactor: got rid of dashboard 1 and dashboard 2
