+++ client/src/Pages/AddEvent/AddEvent.js
@@ -1,0 +1,28 @@
+import './AddEvent.scss';
+import { useState } from 'react';
+import Navbar from '../../Components/Navbar';
+import EventForm from '../../Components/EventForm';
+import { postEventDB } from '../../services/APIservice';
+
+
+function AddEvent () {
+
+  const [event, setEvent] = useState([])
+
+
+  async function postEvent (newEvent) {
+    const createEvent = await postEventDB(newEvent);
+    setEvent(prevState => [...prevState, createEvent]);
+  }
+
+  return (
+    <div>
+      <Navbar />
+      <div className='dashboard-container2'>
+        <EventForm postEvent={postEvent}/>
+      </div>
+    </div>
+  )
+}
+
+export default AddEvent
