--- client/src/Pages/Addevent.js
@@ -1,16 +1,0 @@
-import './Addevent.scss';
-import Navbar from '../Components/Navbar';
-import Dashboard2 from '../Components/Dashboard2';
-
-
-function Addevent () {
-
-  return (
-    <div>
-      <Navbar />
-      <Dashboard2 />
-    </div>
-  )
-}
-
-export default Addevent
