+++ client/src/Pages/MyEvents/MyEvents.scss
